requirements.txt
asgiref==3.9.1
Django==5.2.4
django-role-permissions==3.2.0
djhtml==3.0.9
dotenv==0.9.9
gunicorn==23.0.0
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dotenv==1.1.1
sqlparse==0.5.3
